(#15): Make TripRecyclerFragment not use earlyDawn background.
Previous to this change, the TripRecyclerFragment's layout used the earlyDawn color as a background. This looks fine for cards, but for other layouts it looks a little strange.

Also, because of the floating action button, the card layout looks a little cluttered. As such, I made this duplicate what Hangouts (and other Google apps) do, and just use a white background for the list/recycler view.